+++ src/bes/lims/impress/reportview.py
@@ -622,3 +622,30 @@ class DefaultReportView(SingleReportView):
         """Returns whether the analysis passed-in is out-of-stock
         """
         return api.get_review_status(analysis) == "out_of_stock"
+
+    def is_valid_status(self, analysis):
+        """Returns whether the analysis object or brain is in a valid status
+        """
+        invalid = ["retracted", "rejected", "cancelled"]
+        return api.get_review_status(analysis) not in invalid
+
+    def is_results_in_progress(self, sample):
+        """Returns true if there are analyses not yet verified
+        """
+        def in_progress(analysis):
+            if analysis.getDateVerified():
+                return False
+            if self.is_out_of_stock(analysis):
+                # do not display results as provisional for analyses that are
+                # in the out-of-stock analysis
+                # https://github.com/beyondessential/bes.lims/pull/69
+                return False
+            return True
+
+        # Exclude invalid analyses
+        analyses = self.get_analyses(sample)
+        analyses = filter(self.is_valid_status, analyses)
+
+        # Ensure all valid analyses are in progress
+        analyses = map(in_progress, analyses)
+        return any(analyses)